test(Connection) make sure raised ExecutionErrors are properly handled
