+++ spec/support/star_wars/schema.rb
@@ -96,8 +96,12 @@ module StarWars
           case args[:nameIncludes]
           when "error"
             all_ships = GraphQL::ExecutionError.new("error from within connection")
+          when "raisedError"
+            raise GraphQL::ExecutionError.new("error raised from within connection")
           when "lazyError"
             all_ships = LazyWrapper.new { GraphQL::ExecutionError.new("lazy error from within connection") }
+          when "lazyRaisedError"
+            all_ships = LazyWrapper.new { raise GraphQL::ExecutionError.new("lazy raised error from within connection") }
           else
             all_ships = all_ships.select { |ship| ship.name.include?(args[:nameIncludes])}
           end
